# Specialize media type matching for a single range