While looking at code, I found this low hanging fruit. Matching media types is much simpler for a single one than multiple, so specializing the code for a single argument improves performance of matching. I found most callers to be matching against a single format (`GrpcService`, `THttpService`) so it's some free (micro-) performance.

After
```
Benchmark                             Mode  Cnt         Score          Error  Units
MediaTypesBenchmark.all              thrpt    5  10201031.774 ±   334154.181  ops/s
MediaTypesBenchmark.complexMatch     thrpt    5  28234112.782 ±  4971845.081  ops/s
MediaTypesBenchmark.complexNotMatch  thrpt    5  25309947.033 ±  2968303.292  ops/s
MediaTypesBenchmark.simpleMatch      thrpt    5  93640157.481 ± 12225237.586  ops/s
MediaTypesBenchmark.simpleNotMatch   thrpt    5  79549461.487 ± 12319334.619  ops/s
```

Before
```
Benchmark                             Mode  Cnt         Score         Error  Units
MediaTypesBenchmark.all              thrpt    5   7383325.445 ±  204680.181  ops/s
MediaTypesBenchmark.complexMatch     thrpt    5  23564626.484 ±  418400.507  ops/s
MediaTypesBenchmark.complexNotMatch  thrpt    5  21315281.225 ± 2725154.281  ops/s
MediaTypesBenchmark.simpleMatch      thrpt    5  46150165.428 ± 3222182.395  ops/s
MediaTypesBenchmark.simpleNotMatch   thrpt    5  54324601.873 ± 1090724.032  ops/s
```